% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callIndel.R
\name{callIndel}
\alias{callIndel}
\title{Call insertion/deletion variants}
\usage{
callIndel(bam, genome, genomedir, SNPlocs, threads, outputdir = getwd())
}
\arguments{
\item{bam}{A list of paths to BAM files}

\item{genome}{A BSgenome object, a GmapGenome object, or a character string indicating the reference genome eg. "hg19"}

\item{genomedir}{The directory containing the reference genome or, if genome is a character string, the parent directory of the reference genome directory.}

\item{SNPlocs}{An SNPlocs object containing dbSNP IDs. For indels, this may be an XtraSNPlocs object.}

\item{threads}{The number of cores to use. Cores should equal a factor of reference genome sequence levels ie. chromosome contigs should be equally divisible between cores.}

\item{outputdir}{The output directory for the VCF file.}
}
\value{
A list of file paths to the VCF files.
}
\description{
This function calls indel variants from BAM files. The results are formatted
into a VCF file and the ID column is populated with dbSNP IDs.
}
