% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSNP.R
\name{callSNP}
\alias{callSNP}
\title{Call single nucleotide polymorphism variants}
\usage{
callSNP(bam, genome, genomedir, SNPlocs, threads = 4L, outputdir = getwd())
}
\arguments{
\item{bam}{A list of paths to BAM files}

\item{genome}{A BSgenome object, a GmapGenome object, or a character string indicating the reference genome eg. "hg19"}

\item{genomedir}{The directory containing the reference genome or, if genome is a character string, the parent directory of the reference genome directory.}

\item{SNPlocs}{An SNPlocs object containing dbSNP IDs.}

\item{threads}{The number of cores to use.}

\item{outputdir}{The output directory for the VCF file.}
}
\value{
A list of file paths to the VCF files.
}
\description{
This function calls SNP variants from BAM files. The results are formatted
into a VCF file and the ID column is populated with dbSNP IDs.
}
