% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqProcession.R
\name{fastqProcession}
\alias{fastqProcession}
\title{Convert a FASTQ file to an indexed BAM}
\usage{
fastqProcession(
  file = list_files_with_exts(dir = dir, exts = "fastq"),
  dir = getwd(),
  genome,
  genomedir,
  paired = FALSE,
  threads = 4L,
  molecule = "RNA"
)
}
\arguments{
\item{file}{A list of paths to FASTQ files. If no paths are entered, defaults to all fastq files in dir.}

\item{dir}{Output directory.}

\item{genome}{A BSgenome object, GmapGenome object, or a character string indicating the genome name eg. "hg19".}

\item{genomedir}{A directory containing the reference genome. Otherwise, it is the parent directory of the reference genome is genome is a character string or BSgenome object.}

\item{paired}{Indicates whether the samples are from paired-end or single-end reads.}

\item{threads}{The number of cores to use in the process.}

\item{molecule}{A character string indicating either DNA or RNA samples.}
}
\value{
A list of file paths to created BAM files
}
\description{
This function takes in FASTQ files and performs quality control before
aligning to a reference genome. It assumes paired-end samples are of the same
file name with an underscore (_) and a number to signify different reads of
the same sample. Each sample's outputs will be stored in a separate
directory.
}
