% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_Counting.R
\name{gene_Counting}
\alias{gene_Counting}
\title{Count genes}
\usage{
gene_Counting(
  dir = getwd(),
  groups,
  TxDb,
  orgDb,
  outputdir = getwd(),
  threads = 4L,
  paired = FALSE
)
}
\arguments{
\item{dir}{The parent directory of the sample groups.}

\item{groups}{Folder names of the sample groups. The default is all folders in dir.}

\item{TxDb}{A TxDb object upon which regions of the genome are counted.}

\item{orgDb}{An orgDb object for annotating the CSV with gene symbols and Ensembl IDs.}

\item{outputdir}{Output directory of CSV file.}

\item{threads}{Number of cores to use.}

\item{paired}{Indicates whether the samples are from paired-end reads.}
}
\value{
A data frame containing gene counts.
}
\description{
This function counts reads of gene regions between sample groups.
It assumes that sample BAM files are ordered in a directory structure such as
"group/sample/" as fastqProcession() would order it. It outputs a CSV file
showing gene counts. Works similarly to geneExpression(), but outputs count
data instead of differential expression data.
}
